Fixing bug #2090834 related to custom tag in content manager.
The content of the page is sanitized using MyTextSanitizer::previewTarea instead of displayTarea. This will need to be checked in next release to see why we are not using the correct method and why anyway we are sanitizing the ouput as this is something that should be dealt with at the XoopsObject level anyway (XoopsObject->getVar() is there to take care of this).

Anyway, in order to change the code as less as possible within 1.1 release cycle, I have added a beforePreviewTarea preload event and added the corresonding event in customtag preload. This solves the problem.