removing HTML.Trusted and is admin from class.
HTML.Trusted is useless under the core we have at moment because we are either filtering on input (which is best method), filtering on output, or filtering both input/output.

now with html.trusted, it allows the admin to post tags that can be dangerous or not allowed, but being as some places filter on output, it's pointless because those tags etc will be filtered out for non admin users, so that will mess with the post.  if we were just filtering on input, we could make use of it. but not until.