@stranger: agreed.
I'm adding a check to see if allow_url_fopen is On and if Off, we will use Curl to do so. However, this needs testing and completion.

Also, I'm remving @author tags as I believe we should not continue to put this. We are a team, not individuals. More can be found here:
http://community.impresscms.org/modules/newbb/viewtopic.php?topic_id=2597&post_id=24356#forumpost24356