We have a few problem with TinyMCE when enabled by default. One of the cause is that some modules don't play nice with WYSIWYG editors. A quick way to fix this is only giving WYSIWYG group permission for modules we know it works well.
But this is currently not possible for Webmasters because in XOOPS history, Webmaters always have access to everything. XoopsGroupPermission::checkRight() always return true if if you are Webmaster. 

I have added a 5th parameter in this method, $webmasterAlwaysTrue which is "true" by defaut. So by unless you specify this 5th parameter as "false" absolutely nothing will change, the actual way will be preserve.

But if you specify webmasterAlwaysTrue to "false", the checkRight() method will actually do a real verification and return false if no corresponding permission was found for this user.

So, with this change, I was able to introduce WYSIWYG permissions checkbox for Webmasters group. You can now specify in which module the Webmasters can or cannot use WYSIWYG editor.

Please everyone test this. We are 1 day to final release.

Thanks !