changing session again.
instead of regenerating session on every page (which is causing issues with admin for some users). we regenerate the id upon successful user login. the id will stay constant throughout the logged in session.  if a user is admin, and they goto admin section then the session id is again regenerated and the old one destroyed, but the session id will then be current for the session. if an admin returns to the userside, a new session id will again be regenerated..