fixes for issues in > http://community.impresscms.org/modules/newbb/viewtopic.php?topic_id=2977&forum=63&post_id=27704#forumpost27704
removed session regenerate for admin section, to fix issue of multiple clicks logging some people out & issue with opening pages in multiple browser tabs, whilst working in admin and still browsing on userside in another browser tab. (the session was constantly regenerated in that situation). but the issue of logging you out when clicking multiple times (or repeatedly refreshing) is more an issue of PHP and not the ICMS core, according to references on php.net.

will have a better method of session & cookie handling in 1.2