adding new Data Filter classes (incomplete)
contains functions for handling text & HTML split into 3.

kernel/icms_DataFilter.php - this is the handler class for data filtering, and will select the correct filter (input/output) when required.
class/icms_InputFilter.php - this class is an extension of icms_DataFilter and will be used for validating & sanitizing of User INPUT (Prior to submitting to DB)
class/icms_OutputFilter.php - this class is an extension of icms_DataFilter and will be used for filtering content before rendering by the browser, but after it has been retrieved from the DB.