File : install/page_modulesinstall.php : Message: Notice: Use of undefined constant _CORE_PAGENOTDISPLAYED - assumed '_CORE_PAGENOTDISPLAYED' (This page cannot be displayed due to an internal error etc. etc.) hander / class cannot be found problem. Handler is config handler.
The problem seems to be fixed now.
The new autoload function wasn't in the /install/common.php file
so the config class file (in this case) wasn't loaded automatically.

Therefore no config handler was present.

In the /include/functions.php the including of the config class file was commented out
so therefore the error occurred.

I added the autoload function in /install/common.php and it seems to work
(debug is off in /install/common.php)

Now it installs properly (the notice of missing constant can be fixed later)