Completion of the conversion of the XoopsTpl class to icms_view_Tpl class
Please note: the 2 functions that were moved into the icms_view_Tpl class as static functions can only be called statically when using the full class name and not a variable - this isn't supported until PHP5.3.0 and we are not making that a requirement for this version.

So, throughout the core, $icmsAdminTpl->template_touch($tplid) instead of $icmsAdminTpl::template_touch($tplid) is used to support older versions of PHP5.