Improvements to the database character set and collation settings. Test #489
A few things were happening here - the list of collations wasn't correct because of the use of a wildcard match (collations start with the charset, but there is utf8 and utf8mb4). Used a more specific query to request the associated collations for a charset.

Not all db users will have the ability to modify an existing database. When the ALTER DATABASE statement fails to change the charset and collation of an existing db, we should gracefully accept the existing charset and collation.