Added default iframe settings and domains to HTML Purifier, will work for a fresh install and upgrades from previous versions. Test #458
Using modules/system/include/update.php may seem a bit unorthodox for setting values during a fresh install, but it is preferred for a few important reasons-
* It reduces code duplication
* It reduces code errors

This method allows us to be much more accurate in our releases and we are better assured that an upgraded site will have the same features and settings as a freshly installed site.

I also corrected a logic branch in the icms_config_item_Handler that would not return properly if a textsarea element was an array and not a string.